+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix duplicate entries showing up
 - Show ammo groups under a type in a table by default
 - Only show historical ammo type information when displaying "Show used" in table
+- Only show historical ammo group information when displaying "Show used" in table
 
 # v0.8.5
 - Add link in readme to github mirror